Moving Iris Classification class to Ax Folder
