+++ examples/pytorch/AxHyperOptimizationPTL/AxHyperOptimizationPTL.py
@@ -1,13 +1,8 @@
-import sys, os
-
-current_file = os.path.dirname(os.path.realpath(__file__))
-sys.path.insert(1, current_file.split("examples")[0])
-
 import argparse
 import mlflow
 from ax.service.ax_client import AxClient
-from tests.pytorch.iris import IrisClassification
-from tests.pytorch.iris_data_module import IrisDataModule
+from iris import IrisClassification
+from iris_data_module import IrisDataModule
 import pytorch_lightning as pl
