+++ tests/pytorch/test_pytorch_autolog.py
@@ -2,13 +2,19 @@ from distutils.version import LooseVersion
 import pytest
 import pytorch_lightning as pl
 import torch
-from iris import IrisClassification, IrisClassificationWithoutValidation
+from examples.pytorch.AxHyperOptimizationPTL.iris import (
+    IrisClassification,
+    IrisClassificationWithoutValidation,
+)
 import mlflow
 import mlflow.pytorch
 from pytorch_lightning.callbacks.early_stopping import EarlyStopping
 from pytorch_lightning.callbacks import ModelCheckpoint
 from mlflow.utils.file_utils import TempDir
-from iris_data_module import IrisDataModule, IrisDataModuleWithoutValidation
+from examples.pytorch.AxHyperOptimizationPTL.iris_data_module import (
+    IrisDataModule,
+    IrisDataModuleWithoutValidation,
+)
 from mlflow.utils.autologging_utils import BatchMetricsLogger
 from mlflow.pytorch._pytorch_autolog import _get_optimizer_name
 from unittest.mock import patch
